requirements.txt
fastapi
openai
uvicorn
python-dotenv
langchain
langchain-community
langchain-core
langchain-openai
langchain-qdrant
langchain-experimental
langfuse==2.53.9
qdrant-client
python-multipart
pypdf
gunicorn
paddlepaddle==3.0.0rc1
pdfminer.six==20231228
pydantic==2.10.6
lxml==5.3.0
pillow==11.1.0
tiktoken==0.8.0
open_clip_torch==2.30.0 
streamlit==1.41.1

boto3>=1.28.43
Brotli>=1.1.0
click>=8.1.7
PyMuPDF>=1.24.9,<=1.24.14
loguru>=0.6.0
numpy>=1.21.6,<2.0.0
fast-langdetect>=0.2.3,<0.3.0
scikit-learn>=1.0.2
unimernet==0.2.3
torch>=2.2.2,<=2.3.1
torchvision>=0.17.2,<=0.18.1
matplotlib
ultralytics>=8.3.48
paddleocr==2.7.3
struct-eqtable==0.3.2
einops
accelerate
rapidocr-paddle>=1.4.5,<2.0.0
rapidocr-onnxruntime>=1.4.4,<2.0.0
rapid-table>=1.0.3,<2.0.0
doclayout-yolo==0.0.2b1
detectron2